fix(arch,solver): repair external-push arch-size + clippy breakage on main
Main went red at 8917a37e from direct pushes that bypassed the arch guard and
clippy. Mechanical, behavior-neutral repairs.

arch-size (5 files over their size ceilings; trims/split, no ratchet bumps):
- query_boundaries/assignability.rs 2004->2000, types/interface_type.rs
  2002->2000: condense doc comments.
- flow/control_flow/core.rs 1854->1848, types/utilities/core.rs 1706->1703:
  condense doc comments to their ratchets.
- types/computation/helpers.rs 2039->1859: extract the interface-member type
  computation (get_type_of_interface_member[_simple], merged_method_signature_type)
  to a new sibling types/computation/interface_member_type.rs. Pure code motion
  (inherent CheckerState methods, same crate); drop the now-unused Visibility
  import from helpers.rs.

clippy (tsz-solver, denied lints landed unchecked):
- narrowing/core.rs: remove the unreachable duplicate Some(TypeData::Mapped) arm
  (the documented #15392 arm at the earlier position already handles it) — clears
  the unreachable-pattern, match-same-arms, and one clone-on-Copy at once; fix the
  remaining clone-on-Copy ((*mapped).clone() -> *mapped).
- evaluation/evaluate/application.rs: move the SameAliasExpansion enum above the
  #[cfg(test)] module (items-after-test-module), and #[allow(dead_code)] the
  DivergentConditional variant — a designed-but-unwired verdict
  (classify_same_alias_expansion never returns it yet; the resolve_application
  match arm already handles it), preserved rather than removed.
